pyproject.toml
[tool.poetry]
name = "aiodocker-images"
version = "0.1.0"
description = "Tool for async pull && save images"
authors = ["1ndef1n1te"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiodocker = "^0.21.0"
aiofiles = "^23.2.1"
pyyaml = "^6.0.1"
pydantic = "^2.8.2"
loguru = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
